cloud-function/requirements.txt
functions-framework==3.5.0
Flask==3.0.0
SQLAlchemy==2.0.23
pyodbc==5.1.0
python-dotenv==1.0.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
